requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==21.4.0
    # via jsonschema
bcrypt==3.2.0
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.12
    # via requests
cryptography==3.4.8
    # via pyjwt
django==4.1
    # via
    #   -r requirements.in
    #   django-axes
    #   django-filter
    #   django-healthchecks
    #   django-storages
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-axes==5.31.0
    # via -r requirements.in
django-behaviors==0.5.1
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-healthchecks==1.4.2
    # via -r requirements.in
django-ipware==3.0.7
    # via
    #   -r requirements.in
    #   django-axes
django-split-settings==1.1.0
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-jwt
    #   drf-spectacular
djangorestframework-camel-case==1.3.0
    # via -r requirements.in
drf-jwt==1.19.2
    # via -r requirements.in
drf-spectacular[sidecar]==0.21.2
    # via -r requirements.in
drf-spectacular-sidecar==2022.3.14
    # via drf-spectacular
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.4.0
    # via drf-spectacular
pillow==9.0.1
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.3.0
    # via drf-jwt
pyrsistent==0.18.1
    # via jsonschema
pytz==2021.3
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.27.1
    # via django-healthchecks
sentry-sdk==1.5.8
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   django-healthchecks
sqlparse==0.4.2
    # via django
tzdata==2022.2
    # via django
uritemplate==3.0.1
    # via drf-spectacular
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
whitenoise==6.0.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
